add isRegistered() function to query helper
